Add some plotting functionality
None of this has been tested (at least by Jacob), so no idea if it works.
I have a sneaking suspicion that it won't work in the biscuiteer framework.

Two scripts have been added to inst/scripts for plotting examples (neither of
which uses the functions created in the R code directory). The rest of the
changes can be found in the R directory (and corresponding man page updates).
A few minor changes were made to fix some BiocCheck complaints as well.