Fix case where starts match but ends don't
When collapsing to fragments, there is an edge case where the starts
matched, but the ends didn't. If end(r2) > end(r1), then this is still
considered a proper pair. However, if end(r2) < end(r1), then we can
treat this in the same way as if end(r2) == end(r1).